refactor(store): store 단위 테스트 전반적인 리팩토링 #21
