+++ src/test/resources/application.yml
@@ -7,17 +7,15 @@ spring:
   datasource:
     driver-class-name: org.h2.Driver
     url: jdbc:h2:mem:test;MODE=PostgreSQL;CASE_INSENSITIVE_IDENTIFIERS=true
-    #    DATABASE_TO_UPPER=false;
     username: sa
     password:
     sql:
       init:
         mode: always
-        schema-locations: classpath:schema.sql
   jpa:
     database: postgresql
     hibernate:
-      ddl-auto: update
+      ddl-auto: create-drop
 
 management:
   endpoints:
